Add proper prone condition handling
If a unit is prone, they lose the ability to boost same as slowed.
Additionally, their movement is halved (or treated as difficult terrain
if terrain ruler is active). No handling of units that ignore difficult
terrain is included. If a unit starts their turn with prone, but removes
it by spending their movement to get up, only show boost movent.