Shim dnd5e data to work around elevationruler bug
Elevation Ruler rolled its own measurement function that does not work
correctly outside of dnd5e. This shims in the data that dnd5e puts into
the grid so that Elevation Ruler actually measures correctly.
Additionally, Elevation Ruler measures wrong for hex grids unless the
diagonal rule is set to equidistant ("555" in dnd parlance, "111" in
lancer.) To work around this, if the grid is hex, always apply the
lancer 111 rule.